# perf(dts-plugin): enable tsbuildinfo