--- a/.changeset/cuddly-turkeys-cry.md
+++ b/.changeset/cuddly-turkeys-cry.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/dts-plugin': minor
+---
+
+feat: enable tsc incremental build