--- a/apps/website-new/docs/zh/guide/troubleshooting/type/TYPE-001.mdx
+++ b/apps/website-new/docs/zh/guide/troubleshooting/type/TYPE-001.mdx
@@ -10,6 +10,8 @@ import ErrorCodeTitle from '@components/ErrorCodeTitle';
 
 在 terminal 中执行报错信息参数中的 `cmd` 命令，根据错误信息对文件或者 `tsconfig` 进行修复。
 
+若你希望忽略 ts 中的类型检查错误，可在 `tsconfig.json` 中设置 [`compilerOptions.noCheck`](https://www.typescriptlang.org/tsconfig/#noCheck) 为 `true`（该选项仅支持 ts 5.5 及以上版本）
+
 若 `cmd` 命令执行无报错，但仍然报错 TS 编译失败，需要检查 `ModuleFederationPlugin` 中的 `exposes`字段：
 
 ```ts title="[modern|rspack|rsbuild|webpack].config.[js,ts]"